// source/frontend/src/models/api/generated/ApiGen_Base_Page.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */

// LINK: @backend/apimodels/Models/Base/PageModel.cs
export interface ApiGen_Base_Page<T> {
  items: T[] | null;
  page: number;
  quantity: number;
  total: number;
}